+++ catsup/parser/__init__.py
@@ -3,7 +3,8 @@ from .markdown import md
 
 
 def markdown(content, **kwargs):
-    return md.render(content, **kwargs)
+    html = md.render(content, **kwargs)
+    return html
 
 
 def config(*args, **kwargs):